requirements.txt
python-dotenv
requests
pandas
supabase 